Multiple updates to the IVector interface
First, get(...) and set(...) methods where treating column vectors as
row vectors and vice-versa

Second, all methods inherited from IMatrix that return IMatrix are
overridden to return IVector instead.

Third, when performing vector operations, the specified vector is
transposed if needed, to match the orientation of the operated on
vector.